[FIX] Last minute fix

+++ book_review/views/views_review.py
@@ -1,3 +1,4 @@
+from django.contrib import messages
 from django.contrib.auth.decorators import login_required
 from django.shortcuts import render, redirect, get_object_or_404
 from django.utils.translation import gettext as _
@@ -7,63 +8,36 @@ from book_review.models import Ticket, Review
 
 
 @login_required()
-def create_review(request, review_id=None):
-    instance_review = Review.objects.get(
-        pk=review_id) if review_id is not None else None
-    instance_ticket = instance_review.ticket if instance_review else None
-    title = _("Modify your review") if instance_review else _("Create your review")
+def post_review(request, ticket_id=None, review_id=None):
+    ticket = get_object_or_404(Ticket, pk=ticket_id) if ticket_id else None
+
+    review = get_object_or_404(Review, pk=review_id, user=request.user) if \
+        review_id else None
+    title = _("Modify your review") if review else _("Create your review")
+
+    ticket_form = TicketForm(
+        request.POST if request.method == "POST" else None,
+        request.FILES if request.method == "POST" else None,
+        instance=ticket)
+    review_form = PostReviewForm(
+        request.POST if request.method == "POST" else None,
+        instance=review
+    )
+
     if request.method == 'GET':
-        ticket_form = TicketForm(instance=instance_ticket)
-        review_form = PostReviewForm(instance=instance_review)
-        return render(request, 'review/create_review.html', {
-            'ticket_form': ticket_form,
-            'review_form': review_form,
-            'main_h1': title
-        })
+        return render(request, 'review/post_review.html', locals())
     elif request.method == 'POST':
-        ticket_form = TicketForm(request.POST, request.FILES, instance=instance_ticket)
-        review_form = PostReviewForm(request.POST, instance=instance_review)
-        if review_form.is_valid() and ticket_form.is_valid():
-            ticket = ticket_form.save(commit=False)
-            ticket.user = request.user
-            ticket.save()
+        if review_form.is_valid():
+            if ticket_form and ticket_form.is_valid():
+                ticket = ticket_form.save(commit=False)
+                ticket.user = request.user
+                ticket.save()
             review = review_form.save(commit=False)
             review.user = request.user
             review.ticket = ticket
             review.save()
-    return redirect('my_posts')
-
-
-@login_required()
-def post_review(request, ticket_id):
-    ticket = get_object_or_404(Ticket, pk=ticket_id)
-    instance_review = Review.objects.get(
-        ticket_id=ticket_id, user_id=request.user.id
-    )
-    title = _("Modify your review") if instance_review else _("Create your review")
-
-    if request.method == 'GET':
-        form = PostReviewForm(instance=instance_review)
-        return render(request, 'review/post_review.html',
-                      {
-                          'form': form,
-                          'ticket': ticket,
-                          'main_h1': title,
-                      })
-    elif request.method == 'POST':
-        form = PostReviewForm(request.POST, instance=instance_review)
-        if form.is_valid():
-            review = form.save(commit=False)
-            review.user = request.user
-            review.ticket = ticket
-            review.save()
         else:
-            return render(request, 'review/post_review.html',
-                          {
-                              'form': form,
-                              'ticket': ticket,
-                              'main_h1': title,
-                          })
+            return render(request, 'review/post_review.html', locals())
     return redirect('feed')
 
 
@@ -74,7 +48,10 @@ def view_review(request, review_id):
 
 
 @login_required()
-def delete_review(request, review_id):
-    review = get_object_or_404(Review, pk=review_id)
-    review.delete()
+def delete_review(request):
+    if request.method == "POST" and request.POST.get("delete_review_id"):
+        review = get_object_or_404(Review, pk=request.POST.get("delete_review_id"))
+        review.delete()
+    else:
+        messages.warning(request, _("Your review has'nt been deleted"))
     return redirect('my_posts')
