+++ book_review/views/views_ticket.py
@@ -1,3 +1,4 @@
+from django.contrib import messages
 from django.utils.translation import gettext as _
 from django.contrib.auth.decorators import login_required
 from django.shortcuts import render, redirect, get_object_or_404
@@ -9,7 +10,7 @@ from book_review.models import Ticket
 @login_required()
 def create_ticket(request, ticket_id=None):
     instance_ticket = Ticket.objects.get(
-        pk=ticket_id) if ticket_id is not None else None
+        pk=ticket_id) if ticket_id else None
     main_h1 = _("Modify your ticket") if instance_ticket else _("Create your ticket")
     if request.method == 'GET':
         form = TicketForm(instance=instance_ticket)
@@ -31,7 +32,10 @@ def view_ticket(request, ticket_id):
 
 
 @login_required()
-def delete_ticket(request, ticket_id):
-    ticket = get_object_or_404(Ticket, pk=ticket_id)
-    ticket.delete()
+def delete_ticket(request):
+    if request.method == "POST" and request.POST.get("delete_ticket_id"):
+        ticket = get_object_or_404(Ticket, pk=request.POST.get("delete_ticket_id"))
+        ticket.delete()
+    else:
+        messages.warning(request, _("Your ticket has'nt been deleted"))
     return redirect('my_posts')
